refactor(lms): SessionManager 의 역할을 Session 이 담당하도록 변경
