+++ STEP2-README.md
@@ -32,6 +32,13 @@
 - [x] 입력된 날짜가 Session 시작일 이전이면 준비중 상태를 반환한다.
 - [x] 입력된 날짜가 Session 시작일, 종료일 사이라면 모집중 상태를 반환한다.
 - [x] 입력된 날짜가 Session 종료일 이후이면 종료 상태를 반환한다.
+- [x] 최대 수강 인원 제한이 없다.
+- [x] 최대 수강 인원 제한이 있다.
+- [x] 자리가 있다면 신청이 가능하다.
+- [x] 자리가 없다면 예외가 발생한다.
+- [x] Session 이 '모집중' 상태가 아니면 신청 시 예외가 발생한다.
+- [x] 수강생이 결제한 금액과 수강료가 일치하지 않으면 예외가 발생한다.
+- [x] 수강생이 결제한 금액과 수강료가 일치하면 신청된다.
 ### DateRange
 - [x] 시작일과 종료일은 필수이다.
 - [x] 종료일은 시작일 이전으로 설정이 불가능하다.
@@ -51,14 +58,6 @@
 ### ImageSize
 - [x] 1MB 를 초과하는 크기가 전달된 경우 예외가 발생한다.
 - [x] 1MB 이하의 크기가 전달된 경우 정상적으로 생성된다.
-### SessionManager
-- [x] 최대 수강 인원 제한이 없다.
-- [x] 최대 수강 인원 제한이 있다.
-- [x] 자리가 있다면 신청이 가능하다.
-- [x] 자리가 없다면 예외가 발생한다.
-- [x] Session 이 '모집중' 상태가 아니면 신청 시 예외가 발생한다.
-- [x] 수강생이 결제한 금액과 수강료가 일치하지 않으면 예외가 발생한다.
-- [x] 수강생이 결제한 금액과 수강료가 일치하면 신청된다.
 ### Capacity
 - [x] 인원 제한이 있다면 true 를 반환한다.
 - [x] 인원 제한이 없다면 false 를 반환한다.
