+++ src/main/java/nextstep/session/domain/Session.java
@@ -1,17 +1,34 @@
 package nextstep.session.domain;
 
+import nextstep.courses.exception.SessionNotRecruitingException;
+import nextstep.users.domain.NsUser;
+
 import java.math.BigInteger;
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Session {
     private final DateRange sessionDateRange;
     private final Money fee;
+    private final Capacity capacity;
+    private final List<NsUser> sessionUserList;
 
     private Session(final DateRange sessionDateRange, final Money fee) {
+        this(sessionDateRange, fee, Capacity.noLimit());
+    }
+
+    private Session(final DateRange sessionDateRange, final Money fee, final Capacity capacity) {
+        this(sessionDateRange, fee, capacity, new ArrayList<>());
+    }
+
+    private Session(final DateRange sessionDateRange, final Money fee, final Capacity capacity, final List<NsUser> sessionUserList) {
         validationSession(sessionDateRange, fee);
 
         this.sessionDateRange = sessionDateRange;
         this.fee = fee;
+        this.capacity = capacity;
+        this.sessionUserList = sessionUserList;
     }
 
     private static void validationSession(final DateRange sessionDateRange, final Money fee) {
@@ -28,8 +45,8 @@ public class Session {
         return new Session(sessionDateRange, Money.of(BigInteger.ZERO));
     }
 
-    public static Session paidSession(final DateRange sessionDateRange, final Money fee) {
-        return new Session(sessionDateRange, fee);
+    public static Session paidSession(final DateRange sessionDateRange, final Money fee, final Capacity capacity) {
+        return new Session(sessionDateRange, fee, capacity);
     }
 
     public boolean isEqualsFee(final Money fee) {
@@ -50,5 +67,33 @@ public class Session {
         }
 
         return SessionStatus.종료;
+    }
+
+    public void apply(final LocalDate applyDate, final NsUser sessionUser) {
+        apply(applyDate, sessionUser, Money.ZERO);
+    }
+
+    public void apply(final LocalDate applyDate, final NsUser sessionUser, final Money fee) {
+        if (!isRecruiting(applyDate)) {
+            throw new SessionNotRecruitingException("모집 기간이 아닙니다.");
+        }
+
+        if (!isEqualsFee(fee)) {
+            throw new SessionNotRecruitingException("수강료가 지불한 금액과 일치하지 않습니다.");
+        }
+
+        if (capacity.isFull(sessionUserList.size())) {
+            throw new SessionNotRecruitingException("수강생이 가득찼습니다.");
+        }
+
+        sessionUserList.add(sessionUser);
+    }
+
+    public boolean hasApplied(final NsUser sessionUser) {
+        return sessionUserList.contains(sessionUser);
+    }
+
+    public boolean hasLimit() {
+        return capacity.hasLimit();
     }
 }
